+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Copy a File (Thanks, [@qwertynerd97](https://github.com/qwertynerd97))
 - Course Quiz Extensions
 - List Announcements (Thanks, [@rmanbaird](https://github.com/rmanbaird))
+- Grade/Comment on Multiple Submissions (Thanks, [@rmanbaird](https://github.com/rmanbaird))
 - Quiz Extensions
 
 ### General